Implement patient and doctor services and corresponding tests.

+++ MedicalDocumentationManager.Persistence/Commands/Address/DeleteAddressCommandHandler.cs
@@ -1,8 +1,9 @@
-﻿using MedicalDocumentationManager.Database.Contexts.Abstractions;
+﻿using MediatR;
+using MedicalDocumentationManager.Database.Contexts.Abstractions;
 
 namespace MedicalDocumentationManager.Persistence.Commands.Address;
 
-public sealed class DeleteAddressCommandHandler
+public sealed class DeleteAddressCommandHandler : IRequestHandler<DeleteAddressCommand>
 {
     private readonly IMedicalDocumentationManagerDbContext _context;
 
@@ -13,15 +14,9 @@ public sealed class DeleteAddressCommandHandler
 
     public async Task Handle(DeleteAddressCommand command, CancellationToken cancellationToken)
     {
-        if (command is null)
-        {
-            throw new ArgumentNullException(nameof(command));
-        }
+        if (command is null) throw new ArgumentNullException(nameof(command));
 
         var addressEntity = await _context.AddressEntities.FindAsync(command.Id, cancellationToken);
-        if (addressEntity!= null)
-        {
-            _context.AddressEntities.Remove(addressEntity);
-        }
+        if (addressEntity != null) _context.AddressEntities.Remove(addressEntity);
     }
 }
